requirements.txt
fastapi
uvicorn
playwright
supabase
pandas
python-dotenv